+++ app/src/components/History/TransactionIcons/Fail.tsx
