+++ app/src/components/History/TransactionIcons/Pending.tsx
